test: split up test-runner-mock-timers test
PR-URL: https://github.com/nodejs/node/pull/55506
